refactor(commands): use Laravel Signature attributes and traits
- Replace AbstractCommand and AbstractPublish base classes with reusable traits
- Declare command signatures and options via #[Signature] and #[Description] attributes
- Remove constructor-based InputOption registration for cleaner Laravel 11+ style
